Fix over-broad link deletion in 'customize' rake task

+++ lib/tasks/customize.rake
@@ -24,13 +24,15 @@ task :customize, :path_list do |t, args|
   # first, clear out all symbolic links from any previous
   # configurations--we don't want to leave any old ones lying around
   # pointing into customization directories we aren't currently using.
-  FileList[ "*", "**/*"].
-    exclude do |maybe_not_link|
-      !File.symlink?(maybe_not_link)
-    end.
-    each do |link_to_delete|
-      File.delete link_to_delete
-    end
+  unless RAILS_ENV == 'production'
+    FileList[ "*", "**/*"].
+      exclude do |maybe_not_link|
+        !File.symlink?(maybe_not_link)
+      end.
+      each do |link_to_delete|
+        File.delete link_to_delete
+      end
+  end
 
   # now, build new links
   paths = args[:path_list].split(':')
